Moved Kieker’s instrumentation classes into their own package.
