+++ Kieker Measurement Tool/src/main/java/de/uka/ipd/sdq/beagle/measurement/kieker/instrumentation/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Instrumentation for the Kieker Measurement Tool.
+ */
+package de.uka.ipd.sdq.beagle.measurement.kieker.instrumentation;
